Fix issue 163 (#168)
* Add default profile picture variable

* Add property that checks for default profile pic

* Test property 'has_default_picture'

* Remove profile picture only if not default

* Check remove conditions on single line

+++ src/safm_api/views.py
@@ -361,7 +361,7 @@ class UserProfileView(generics.GenericAPIView):
                     serializer.is_valid(raise_exception=True)
     
                     # Removes the old profile picture if a new one is uploaded
-                    if request.data.get('profile_picture'):
+                    if request.data.get('profile_picture') and not profile.has_default_picture:
                         path_old_picture = os.path.join(settings.MEDIA_ROOT, profile.profile_picture.name)
                         os.remove(path_old_picture)
